add practice questions

+++ src/banking_app/Transaction.java
@@ -1,7 +1,7 @@
 package banking_app;
 
 public class Transaction {
-    public void transfer(Account fromAccount, Account toAccount, double amount) {
+    public void transfer(Account fromAccount, Account toAccount, int amount) {
         if (fromAccount.withdraw(amount)) {
             toAccount.deposit(amount);
             System.out.println("Transferred " + amount + " from " + fromAccount.getAccountNumber() +
